fixup! Move pixel jittering into camera

+++ src/InOneWeekend/scene.h
@@ -28,12 +28,12 @@ class scene {
 
         for (int j = 0; j < image_height; ++j) {
             std::clog << "\rScanlines remaining: " << (image_height - j) << ' ' << std::flush;
-            auto t = j / (image_height-1);
+            auto t = static_cast<double>(j) / (image_height-1);
 
             for (int i = 0; i < image_width; ++i) {
                 color pixel_color(0,0,0);
                 for (int sample = 0; sample < samples_per_pixel; ++sample) {
-                    auto s = i / (image_width-1);
+                    auto s = static_cast<double>(i) / (image_width-1);
                     ray r = cam.get_ray(s, t);
                     pixel_color += ray_color(r, max_depth);
                 }
